Remove sprockets and use webpacker for all assets
Rails 6 has this weird thing where it includes both, sprockets and webpacker, and only the javascript is handled by webpacker, and the rest continues to go through sprockets.

This is confusing and was giving me lots of trouble when precompiling assets for use in a container in production.

Cleanup all sprockets stuff and setup webpacker to use it not only for javascript but also styles, images, etc.