Add manual deploy to prod workflow
The idea is to have an action to manually trigger a deploy to production (by pulling the SHA and tagging as production.latest and push to ECR).

This workflow should wait for the previous one to pass `CI and CD` and wait for approval before triggering (because of the `environment: production`).

Might require a few tweaks here and there until it is working as intended tho.