Tweak the inputs styling
This will take care of giving the text input and the button a more "govuk" appearance, particularly important for the hover/focus states.

In order to do this there is a need to remove a pesky inline styling that gets precedence no matter what we do, so I use a bit of javascript to "listen" for any DOM changes and remove the `style` attribute, allowing our custom CSS to take place.