Reduce screen content jumping around
When loading or reloading the page, the search would take a few milliseconds to appear while the rest of the screen had been already rendered, so when it appears it jumps the content (common search terms block) down and is briefly noticeable.

Making the screen synchronous reduce/remove the jumping.