Track common terms, an search events
This will track clicks on the common search terms, as well as search events, but not what the user is searching for, just an event to indicate it triggered a search.

Also included jQuery to be easier to reuse some of the code we already had, although it is quite different now.