Mitigates text input jumping on page load
There is some javascript magic that we use to re-style the text input and button on page load. However there is a small delay until this happen and the element made a brief jump on the page.

Setting a min-height ensures the element stays in its place and the common terms (below) will not take its space.