Add more cukes and screenshot diff
Ticket: https://trello.com/c/Naz6YAH0

Follow-up to PR #39. Adding some more cucumber tests, and also a (hopefully) useful thingy to screenshot how the page looks after performing a search, and compare against the previous version, to detect changes.

This should be useful as it is very difficult to create tests for the markup and the styles, but with a screenshot we can catch some regressions more easily.