Disable failures on screenshot diffs when running on CI
Because on CI the browser tests will run on a different machine and setup (Ubuntu), the screenshots are slightly different to those created on local dev machines (Mac).

To not overcomplicate this, I'm disabling the diff check on CI but still archiving the artifacts (the screenshots) for later reference so we can manually inspect them to see if there are any regressions.