Add cookies controller
We remove the cookies action from the about controller and convert it into a `#show` action in the new cookies controller.

In this new controller we also add an `#update` action that will handle the accept/reject buttons from the cookie banner to set the cookie and redirect back.

The cookie will expire after 1 year. This is configurable.