Add the cookies banner and JS to handle it
The cookies banner has 2 parts.

One is the accept/reject banner, and once any of these action buttons have been clicked, and the cookie is set, it redirects back to the page where the user was (or the homepage as a fallback) and shows another confirmation banner that can be hidden clicking a button.

The javascript only handles this close action on the client side so we don't have to trigger another server side request. Unless JS is disabled, in which case we just reload the page so the banner is gone.